% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{function that will output the population over time of a country inside the United Nation from 1950-2020}
\usage{
CountryPopulation(Name)
}
\arguments{
\item{Name}{The name of the county with inside "" with spaces if needed}
}
\value{
Graph showing the population of that country over time
}
\description{
function that will output the population over time of a country inside the United Nation from 1950-2020
}
\examples{
CountryPopulation("United States of America")
}
