% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{Population data of every country in the United Nations from 1950 to 2020}
\format{
A data frame with 71 observations of 236 Variables
\describe{
  \item{Year}{Each year from 1950 to 2020}
  \item{Rest}{Collumn for each country}
}
}
\source{
\url{https://population.un.org/dataportal/}
}
\usage{
WorldPopulation
}
\description{
Population data of every country in the United Nations from 1950 to 2020
}
\keyword{datasets}
